requirements.txt
streamlit>=1.24.0
mysql-connector-python>=8.0.33
python-dotenv>=1.0.0
openai>=1.3.0
anthropic>=0.3.0
langchain-openai>=0.0.3
langchain-anthropic>=0.0.3
langchain-google-genai>=0.0.3
pandas>=2.0.3
numpy>=1.24.3
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
langchain
langchain-experimental
wikipedia
duckduckgo-search
python-docx
PyPDF2
openpyxl
requests
beautifulsoup4
torch
yt-dlp
google-generativeai>=0.3.0

# PDF 파싱을 위한 추가 라이브러리
unstructured
pdfminer.six
PyMuPDF
pymupdf
pdfplumber

# OCR을 위한 라이브러리
pytesseract
Pillow
